+++ test/core/gen2/test_generator.py
@@ -6,13 +6,11 @@ import xarray as xr
 import yaml
 
 from xcube.core.dsio import rimraf
-from xcube.core.gen2.config import CubeGeneratorConfig
 from xcube.core.gen2.generator import CubeGenerator
 from xcube.core.gen2.generator import LocalCubeGenerator
+from xcube.core.gen2.request import CubeGeneratorRequest
 from xcube.core.new import new_cube
 from xcube.core.store.stores.memory import MemoryDataStore
-
-
 
 
 class LocalCubeGeneratorTest(unittest.TestCase):
@@ -46,7 +44,7 @@ class LocalCubeGeneratorTest(unittest.TestCase):
     @requests_mock.Mocker()
     def test_dict(self, m):
         m.put('https://xcube-gen.test/api/v1/jobs/tomtom/iamajob/callback', json={})
-        LocalCubeGenerator(CubeGeneratorConfig.from_dict(self.REQUEST), verbose=True).generate_cube()
+        LocalCubeGenerator(CubeGeneratorRequest.from_dict(self.REQUEST), verbose=True).generate_cube()
         self.assertIsInstance(MemoryDataStore.get_global_data_dict().get('CHL'),
                               xr.Dataset)
