+++ xcube/cli/gen2.py
@@ -35,8 +35,8 @@ import click
               metavar='SERVICE_CONFIG',
               help='A JSON or YAML file that provides the configuration for an'
                    ' xcube Generator Service. If provided, the REQUEST will be'
-                   ' passed to the service instead of generating it on this'
-                   ' computer.')
+                   ' passed to the service instead of generating the cube on '
+                   ' this computer.')
 @click.option('--info', '-i',
               is_flag=True,
               help='Output information about the data cube to be generated.'
@@ -52,19 +52,23 @@ def gen2(request_path: str,
     """
     Generator tool for data cubes.
 
-    Creates a cube view from one or more cube stores, optionally performs some cube transformation,
-    and writes the resulting cube to some target cube store.
+    Creates a cube view from one or more cube stores, optionally performs some
+    cube transformation, and writes the resulting cube to some target cube
+    store.
 
-    REQUEST is the cube generation request. It may be provided as a JSON or YAML file
-    (file extensions ".json" or ".yaml"). If the REQUEST file argument is omitted, it is expected that
-    the Cube generation request is piped as a JSON string.
+    REQUEST is the cube generation request. It may be provided as a JSON or
+    YAML file (file extensions ".json" or ".yaml"). If the REQUEST file
+    argument is omitted, it is expected that the Cube generation request is
+    piped as a JSON string.
 
-    STORE_CONFIGS is a path to a JSON or YAML file with xcube data store configurations.
-    It is a mapping of arbitrary store names to configured data stores. Entries are dictionaries
-    that have a mandatory "store_id" property which is a name of a registered xcube data store.
-    The optional "store_params" property may define data store specific parameters.
-    The following example defines a data store named "my_s3_store" which is an AWS S3 bucket store,
-    and a data store "my_test_store" for testing, which is an in-memory data store:
+    STORE_CONFIGS is a path to a JSON or YAML file with xcube data store
+    configurations. It is a mapping of arbitrary store names to configured
+    data stores. Entries are dictionaries that have a mandatory "store_id"
+    property which is a name of a registered xcube data store.
+    The optional "store_params" property may define data store specific
+    parameters. The following example defines a data store named
+    "my_s3_store" which is an AWS S3 bucket store, and a data store
+    "my_test_store" for testing, which is an in-memory data store:
 
     \b
     {
@@ -94,6 +98,7 @@ def gen2(request_path: str,
     """
     from xcube.core.gen2 import CubeGenerator
     from xcube.core.gen2 import CubeGeneratorError
+    from xcube.core.gen2 import CubeInfo
     from xcube.core.store import DataStoreError
     try:
         generator = CubeGenerator.from_file(request_path,
@@ -101,10 +106,12 @@ def gen2(request_path: str,
                                             service_config_path=service_config_path,
                                             verbose=verbose)
         if info:
-            import sys
-            import yaml
-            cube_info = generator.get_cube_info()
-            yaml.dump(cube_info.to_dict(), stream=sys.stdout, indent=2)
+            def dump_cube_info(cube_info: CubeInfo):
+                import sys
+                import yaml
+                yaml.dump(cube_info.to_dict(), stream=sys.stdout, indent=2)
+
+            dump_cube_info(generator.get_cube_info())
         else:
             generator.generate_cube()
